refactor(ui): extract chat/useLlmModelSelection hook from App.tsx
Moves the /v1/models fetch, the model/reasoning picks with their cookie and
per-session PATCH persistence, the opened-session selection restore, and the
derived maxContextTokens/multimodal/reasoning-levels memos into one hook.
App keeps calling setOpenSessionSelection from loadMessages, resetForNewChat
from goHome, and bumpModelsEpoch from Settings onConfigSaved. Pure move.
